feat(protocol): extend MsgType for rich message types + add validation
- Add 6 new MsgType values: kLocation, kEmoji, kCard, kSystem, kCustom + extension range (>=100)
- Add IsValidMsgType() helper and server-side validation in msg_service
- Add kInvalidMsgType error code (2010)
- Change LastMsgBrief.msg_type from int32_t to MsgType enum for type safety
- Fix implicit MsgType->int conversions in client SDK and conv_service
- Update login same-device-id tests to match reject-new-login semantics

+++ client/nova_sdk/service/conversation_service.cpp
@@ -29,7 +29,7 @@ void ConversationService::GetConversationList(ConvListCallback cb) {
                             .sender_uid      = c.last_msg.sender_uid,
                             .sender_nickname = c.last_msg.sender_nickname,
                             .content         = c.last_msg.content,
-                            .msg_type        = c.last_msg.msg_type,
+                            .msg_type        = static_cast<int>(c.last_msg.msg_type),
                             .server_time     = c.last_msg.server_time,
                         },
                         .mute       = c.mute,
